refactor!: Remove deprecated code.
